backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.1.1
Flask-Compress==1.13
requests==2.28.1
beautifulsoup4==4.12.0
python-dotenv==1.0.0
pinecone-client==2.1.2
google-generativeai==0.1.1
